requirements.txt
fastapi
uvicorn[standard]
python-multipart
openai>=1.30.0
Pillow
python-dotenv
diffusers[torch]
torch
accelerate
transformers